Change default exercise author to hostname
This is a fix for issue #159.

Previously, if the administrator did not specify an author when
creating a new exercise, the author would default to ‘wger.de’. Now, if
the administrator does not specify an author for a new exercise, the
author will default to the hostname that WGER is installed on.

This commit also updates the test to ensure this is what we test for.